Re-tag 1.5 to 2.0rc1
Before I started adding commits to this project, I created a rpmdeplint-1.5 tag
because the previous was rpmdeplint-1.4.

After I released 1.6 I found out that there's a problem which looks like
https://bugzilla.redhat.com/show_bug.cgi?id=1343247#c18
and I realized that the next release after 1.4 was supposed to be 2.0, not 1.5
probably because of the
- hawkey to libsolv move (36fc40c3 and others)
- repodata caching/reusing (6f11c370 and others)
- rearrange conflict checking algorithm (5c07a05d)

So I retroactively replaced rpmdeplint-1.5 with v2.0rc1
and rpmdeplint-1.6 with v2.0rc2.
Next tag will be 2.0rc3