conflict checker: skip packages obsoleted by packages under test (#20)
As explained in issue #20, when checking for conflicts, it's not
enough to skip only existing packages with the same name as the
one they are being checked against. We should skip all existing
packages that the set of packages being tested would replace.

Fortunately, we already have `_select_obsoleted_by` to do this
exact job, so let's just use it here.

We keep the existing 'skip if the names are the same' check to
avoid generating confusing results if the update under test
contains a package with the same name as an existing package but
with a *lower* EVR. This is wrong, but we should leave it to be
caught by other checks; handling it here would give a rather
confusing failure about file conflicts between the two packages,
which the packager might not understand.

Signed-off-by: Adam Williamson <awilliam@redhat.com>